feat: tidy code
