+++ key.go
@@ -18,7 +18,6 @@ import "C"
 
 import (
 	"errors"
-	"log"
 	"math/rand"
 	"reflect"
 	"runtime"
@@ -437,8 +436,9 @@ func keyToggles(k string, keyArr []string, pid int, args ...interface{}) error {
 		return err
 	}
 
-	log.Println("===============debug====================1111112222223333333333333333")
-	C.toggleKeyCode(key, C.bool(down), flags, C.uintptr(pid))
+	C.toggleKeyCode(key, C.bool(down), 
+	
+	flags, C.uintptr(pid))
 	if len(args) > 0 {
 		MilliSleep(KeySleep)
 	}
@@ -544,7 +544,6 @@ func KeyTap(key string, args ...interface{}) error {
 //	robotgo.KeyToggle("a", "up", "alt", "cmd")
 //	robotgo.KeyToggle("k", pid int)
 func KeyToggle(key string, args ...interface{}) error {
-
 	if len(key) > 0 && unicode.IsUpper([]rune(key)[0]) {
 		args = append(args, "shift")
 	}
@@ -668,7 +667,7 @@ func inputUTF(str string) {
 //	robotgo.TypeStr("abc@123, Hi galaxy, こんにちは")
 //	robotgo.TypeStr("To be or not to be, this is questions.", pid int)
 func TypeStr(str string, args ...int) {
-	var tm, tm1 = 0, 7
+	tm, tm1 := 0, 7
 
 	if len(args) > 1 {
 		tm = args[1]
